Allow the srtp modifier to encrypt and decrypt several SRTP streams. It is used for RTP bundle mode with DTLS-SRTP. Fix strdup_printf warning.

+++ tester/mediastreamer2_tester.c
@@ -143,7 +143,7 @@ int main (int argc, char *argv[]) {
 	// this allows to launch tester from outside of tester directory
 	if (strstr(argv[0], ".libs")) {
 		long prefix_length =(long) (strstr(argv[0], ".libs") - argv[0] + 1);
-		char *prefix = ms_strdup_printf("%s%.*s", argv[0][0] == '/' ? "" : "./", prefix_length, argv[0]);
+		char *prefix = ms_strdup_printf("%s%.*s", argv[0][0] == '/' ? "" : "./", (int)prefix_length, argv[0]);
 		ms_warning("Resource prefix set to %s", prefix);
 		bc_tester_set_resource_dir_prefix(prefix);
 		bc_tester_set_writable_dir_prefix(prefix);
